# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Reparto API",
    "version": "0.2.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/v1/usuarios": {
      "get": {
        "summary": "Listar usuarios",
        "tags": [
          "usuarios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usuarios": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string"
                          },
                          "nombre": {
                            "type": "string"
                          },
                          "rol": {
                            "type": "string",
                            "enum": [
                              "admin",
                              "despachador",
                              "chofer"
                            ]
                          },
                          "activo": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "id",
                          "username",
                          "nombre",
                          "rol",
                          "activo"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "usuarios"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear usuario (el nombre de usuario se genera: inicial + primer apellido)",
        "tags": [
          "usuarios"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 60
                  },
                  "apellidoPaterno": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 60
                  },
                  "apellidoMaterno": {
                    "type": "string",
                    "maxLength": 60
                  },
                  "rol": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "despachador",
                      "chofer"
                    ]
                  },
                  "pin": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64
                  }
                },
                "required": [
                  "nombre",
                  "apellidoPaterno",
                  "rol",
                  "pin"
                ]
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    },
                    "nombre": {
                      "type": "string"
                    },
                    "rol": {
                      "type": "string",
                      "enum": [
                        "admin",
                        "despachador",
                        "chofer"
                      ]
                    },
                    "activo": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "username",
                    "nombre",
                    "rol",
                    "activo"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/usuarios/{id}/pin": {
      "post": {
        "summary": "Resetear la clave",
        "tags": [
          "usuarios"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pin": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64
                  }
                },
                "required": [
                  "pin"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true,
                  "enum": [
                    null
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}